chore: feature-home3 모듈 생성 후 파일 이동
