--- feature-home/src/main/kotlin/com/materip/matetrip/ui/HomeScreen.kt
@@ -1,8 +1,0 @@
-package com.materip.matetrip.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun HomeScreen() {
-    // HomeScreen implementation
-}
