+++ feature-home3/src/main/java/com/materip/feature_home/ui/PostScreen.kt
@@ -1,4 +1,4 @@
-package com.materip.matetrip.ui
+package com.materip.feature_home.ui
 
 import androidx.compose.runtime.Composable
