+++ feature-home3/src/main/java/com/materip/feature_home/viewModel/HomeHiltViewModel.kt
@@ -1,4 +1,4 @@
-package com.materip.matetrip.viewModel
+package com.materip.feature_home.viewModel
 
 //@HiltViewModel
 //class HomeHiltViewModel @Inject constructor(
